Move sync features to Turso
